+++ assets/changelog.md
@@ -3,6 +3,7 @@
 ## [0.1.0 build 28] 2025-XX-XX
 ### Changes
 - Made it much harder to do accidental double-submits
+- Updated architecture for providing new questions after receipt of notification
 
 ## [0.1.0 build 28] 2025-05-29
 ### Changes